authserver/src/main/java/com/example/authserver/config/csrf/CsrfCookieFilter.java
package com.example.authserver.config.csrf;

import jakarta.servlet.FilterChain;
import jakarta.servlet.ServletException;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import org.springframework.security.web.csrf.CsrfToken;
import org.springframework.web.filter.OncePerRequestFilter;
import org.springframework.web.servlet.HandlerExceptionResolver;

import java.io.IOException;

public final class CsrfCookieFilter extends OncePerRequestFilter {

    private final HandlerExceptionResolver exceptionResolver;

    public CsrfCookieFilter(HandlerExceptionResolver exceptionResolver) {
        this.exceptionResolver = exceptionResolver;
    }

    @Override
    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) {

        try {
            CsrfToken csrfToken = (CsrfToken) request.getAttribute("_csrf");
            // Render the token value to a cookie by causing the deferred token to be loaded
            csrfToken.getToken();

            filterChain.doFilter(request, response);
        } catch (ServletException | IOException e) {
            this.exceptionResolver.resolveException(request, response, null, e);
        }
    }
}